extract json stuff to own sbt module
